+++ packages/ui/src/components/index.ts
@@ -39,6 +39,9 @@ export {
   InputAdornment,
   Snackbar,
   Radio,
+  Select,
+  InputLabel,
+  Fade,
 } from '@mui/material';
 
 export type {
@@ -80,7 +83,12 @@ export type {
   SlideProps,
   GridProps,
   SnackbarProps,
+  SelectProps,
+  InputLabelProps,
+  FadeProps,
 } from '@mui/material';
+
+export { LoadingButton, type LoadingButtonProps } from '@mui/lab';
 
 export * from './IconButton';
 export * from './InfoTable';
@@ -97,4 +105,5 @@ export * from './CopyButton';
 export * from './OtpInput';
 export * from './Typography';
 export * from './Banner';
+export * from './AmountInput';
 export * from './AddressDropdown';